move code to http_client.cc
